requirements.txt
eth-account==latest
pymongo==latest
web3==latest
python-dotenv==latest